pyproject.toml
[project]
name = "confluence-agent"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["google-adk>=1.17.0", "pip-system-certs>=5.3"]
